Changed how deBruijn indexing works for Tm.
All parameters under a constant are now abstracted over all binders in terms of indexing. Of course, because parameters differ in which of the binders they abstract over, parameters differ in which indices they are allowed to contain while being wellformed.

The big advantage is that binding becomes independent of the KernelContext, and much easier to handle in general.